requirements.txt
pandas
numpy
pandas_ta
lightgbm
optuna
polars
pyyaml
matplotlib
scikit-learn